requirements.txt
selenium<4.10.0
Pillow~=10.0.0
